boards: up_squared_adsp: Update logging frontend
Update frontend moving general library code to appropriate place.
